requirements.txt
# Core ML Libraries
pandas>=2.1.0
numpy>=1.26.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.1

# Visualization
matplotlib>=3.8.0
seaborn>=0.13.0

# Web App
streamlit>=1.28.0

# Google Cloud
google-cloud-aiplatform>=1.35.0
google-cloud-storage>=2.10.0

# Utilities
python-dotenv>=1.0.0
